refactor: move app api handling into dedicated module
